Fixed bug with 'run test' where tests at bottom of file won't run.
The line of code starting with sublime.Region(... was hard-coding a region based on the current position minus 8000 characters. However in situations where the code file has much more than 8000 characters (say 400 lines of test code for example) tests towards the bottom of the file won't run because the resulting text_string will not contain the beginning of the class, which the code will try to find.