Update python_test.py
We want the last match on the class list, not the first. The problem with the previous code is that if you have two test classes in the same file with different names, but the test methods have the same name, then when a single test is run, it would execute the first test found towards to top of the file, not the last test closest the cursor.